add prepared data_maker function from SequenceTagger and IOBTagger to posTagger and chunker
